+++ lib/frozen_record/base.rb
@@ -29,6 +29,7 @@ module FrozenRecord
     FALSY_VALUES = [false, nil, 0, -''].to_set
 
     class_attribute :base_path, :primary_key, :backend, :auto_reloading, :default_attributes, instance_accessor: false
+    class_attribute :clone_on_find, instance_accessor: false
     class_attribute :index_definitions, instance_accessor: false
     class_attribute :max_records_scan, instance_accessor: false
     self.index_definitions = {}.freeze
@@ -261,6 +262,10 @@ module FrozenRecord
 
     def to_key
       [id]
+    end
+
+    def clone
+      self.class.new(attributes)
     end
 
     private